fixed nauf_interaction bug
Before, if you had f1 in a/b/c and f2 in d/e/f, and f2 was always NA
when f1 was a, and f2 could only be e/f and not d when f1 was b, but f2
could be all of d/e/f when f1 was c, nauf_interaction would fail to drop
f2 = d and issue a collinearity warning.  This has been fixed.
Collinearity warnings should only arise when there is an unresolvable
collinearity.